requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.25.1
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
